+++ styles/loginStyle.js
@@ -1,0 +1,18 @@
+import React from "react";
+import { StyleSheet } from "react-native";
+
+const loginStyle = StyleSheet.create({
+  loginView: {
+    borderRadius: 10,
+    height: 55,
+    borderRadius: 6,
+    justifyContent: "center",
+  },
+  loginText: {
+    fontSize: 20,
+    fontWeight: "700",
+    textAlign: "center",
+  },
+});
+
+export default loginStyle;
